refactor: moves labeler.yml related sources to their own folder
