+++ src/main.ts
@@ -3,7 +3,7 @@ import { EOL } from "node:os";
 import { type Writable } from "node:stream";
 import { parseArgs } from "node:util";
 import generateCodeOwners from "./generate-codeowners.js";
-import generateLabelerYml from "./generate-labeler-yml.js";
+import generateLabelerYml from "./labeler-yml/generate.js";
 import readTeamMap from "./read-team-map.js";
 import readVirtualCodeOwners from "./virtual-code-owners/read.js";
 import { VERSION } from "./version.js";
